# src/app/api/v1/openapi.yaml
openapi: 3.0.3
info:
    title: AcidBot API v1
    version: 1.0.0
    description: |
        OpenAPI specification for the routes implemented under `src/app/api/v1`.

        Notes:
        - Several endpoints depend on an authenticated session in the application layer.
        - Image proxy responses return the upstream image bytes with the upstream content type.
        - Date fields are serialized by Next.js/MongoDB to ISO strings in JSON responses.
servers:
    - url: acid.kvznmx.com
      description: Production server
    - url: localhost:3000
      description: Local development server
...
paths:
    /api/v1/profile:
        get:
            tags: [profile]
            summary: Fetch the current maimai profile block
            description: |
                Fetches the maimai mobile profile page using the provided `clal` token,
                then parses the profile block into a normalized JSON object.

                This endpoint requires an authenticated session.
            parameters:
                - in: query
                  name: clal
                  required: true
                  schema:
                      type: string
                  description: The CLAL token used to fetch the maimai profile page.
            responses:
                '200':
                    description: Parsed profile information
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ParsedProfile'
                '401':
                    description: Unauthenticated
                '500':
                    description: Failed to fetch or parse the profile block
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
...
components:
    schemas:
        ErrorResponse:
            type: object
            additionalProperties: true
            required: [error]
            properties:
                error:
                    oneOf:
                        - type: string
                        - type: object
                        - type: array
                        - type: number
                        - type: boolean
        UserCollectionCount:
            type: object
            additionalProperties: false
            required: [img, text]
            properties:
                img:
                    type: string
                    nullable: true
                text:
                    type: string
                    nullable: true
        ParsedProfile:
            type: object
            additionalProperties: false
            required:
                - profilePicture
                - dan
                - rank
                - userName
                - userDetail
                - userCollectionCount
            properties:
                profilePicture:
                    type: string
                    nullable: true
                    description: Same-origin proxied profile image URL when present.
                dan:
                    type: string
                    nullable: true
                    description: Same-origin proxied dan badge image URL when present.
                rank:
                    type: string
                    nullable: true
                    description: Same-origin proxied rank image URL when present.
                userName:
                    type: string
                    nullable: true
                userDetail:
                    type: string
                    nullable: true
                userCollectionCount:
                    nullable: true
                    oneOf:
                        - $ref: '#/components/schemas/UserCollectionCount'
...